+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dto/CertifiedProductSummaryDTO.java
@@ -1,6 +1,7 @@
 package gov.healthit.chpl.dto;
 
 import java.io.Serializable;
+import java.time.LocalDate;
 import java.util.Date;
 
 import gov.healthit.chpl.entity.listing.CertifiedProductDetailsEntity;
@@ -39,6 +40,12 @@ public class CertifiedProductSummaryDTO implements Serializable {
     private Date creationDate;
     private Date lastModifiedDate;
     private String lastModifiedUser;
+    private String rwtPlansUrl;
+    private LocalDate rwtPlansCheckDate;
+    private String rwtResultsUrl;
+    private LocalDate rwtResultsCheckDate;
+    private Integer rwtEligibilityYear;
+
 
     /**
      * Construct from entity.
@@ -54,6 +61,8 @@ public class CertifiedProductSummaryDTO implements Serializable {
         this.acb = new CertificationBodyDTO();
         this.acb.setId(entity.getCertificationBodyId());
         this.acb.setName(entity.getCertificationBodyName());
+        this.acb.setAcbCode(entity.getAcbCode());
+        this.acb.setWebsite(entity.getCertificationBodyWebsite());
         this.developer = new DeveloperDTO();
         this.developer.setName(entity.getDeveloperName());
         ContactDTO contact = new ContactDTO();
@@ -71,6 +80,11 @@ public class CertifiedProductSummaryDTO implements Serializable {
         this.creationDate = entity.getCreationDate();
         this.lastModifiedDate = entity.getLastModifiedDate();
         this.lastModifiedUser = entity.getLastModifiedUser();
+        this.rwtEligibilityYear = entity.getRwtEligibilityYear();
+        this.rwtPlansUrl = entity.getRwtPlansUrl();
+        this.rwtPlansCheckDate = entity.getRwtPlansCheckDate();
+        this.rwtResultsUrl = entity.getRwtResultsUrl();
+        this.rwtResultsCheckDate = entity.getRwtResultsCheckDate();
     }
 
     public CertifiedProductSummaryDTO(CertifiedProductDetailsEntity entity) {
@@ -83,7 +97,10 @@ public class CertifiedProductSummaryDTO implements Serializable {
         this.acb = new CertificationBodyDTO();
         this.acb.setId(entity.getCertificationBodyId());
         this.acb.setName(entity.getCertificationBodyName());
+        this.acb.setAcbCode(entity.getCertificationBodyCode());
+        this.acb.setRetired(entity.getAcbIsRetired());
         this.developer = new DeveloperDTO();
+        this.developer.setId(entity.getDeveloperId());
         this.developer.setName(entity.getDeveloperName());
         ContactDTO contact = new ContactDTO();
         contact.setEmail(entity.getEmail());
@@ -91,13 +108,20 @@ public class CertifiedProductSummaryDTO implements Serializable {
         contact.setFullName(entity.getFullName());
         this.developer.setContact(contact);
         this.product = new ProductDTO();
+        this.product.setId(entity.getProductId());
         this.product.setName(entity.getProductName());
         this.version = new ProductVersionDTO();
+        this.version.setId(entity.getProductVersionId());
         this.version.setVersion(entity.getProductVersion());
         this.reportFileLocation = entity.getReportFileLocation();
         this.sedReportFileLocation = entity.getSedReportFileLocation();
         this.transparencyAttestationUrl = entity.getTransparencyAttestationUrl();
         this.creationDate = entity.getCreationDate();
         this.lastModifiedDate = entity.getLastModifiedDate();
+        this.rwtEligibilityYear = entity.getRwtEligibilityYear();
+        this.rwtPlansUrl = entity.getRwtPlansUrl();
+        this.rwtPlansCheckDate = entity.getRwtPlansCheckDate();
+        this.rwtResultsUrl = entity.getRwtResultsUrl();
+        this.rwtResultsCheckDate = entity.getRwtResultsCheckDate();
     }
 }