+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/questionableactivity/ListingQuestionableActivityProvider.java
@@ -370,7 +370,7 @@ public class ListingQuestionableActivityProvider {
         QuestionableActivityListingDTO activity = null;
         if (origListing.getSurveillance() != null && origListing.getSurveillance().size() > 0
                 && (newListing.getSurveillance() == null
-                        || newListing.getSurveillance().size() < origListing.getSurveillance().size())) {
+                || newListing.getSurveillance().size() < origListing.getSurveillance().size())) {
 
             activity = new QuestionableActivityListingDTO();
             activity.setBefore(null);
@@ -444,6 +444,28 @@ public class ListingQuestionableActivityProvider {
         return activity;
     }
 
+    public QuestionableActivityListingDTO checkRealWorldTestingPlanRemoved(
+            CertifiedProductSearchDetails origListing, CertifiedProductSearchDetails newListing) {
+
+        QuestionableActivityListingDTO activity = null;
+        if (!StringUtils.isEmpty(origListing.getRwtPlansUrl()) && StringUtils.isEmpty(newListing.getRwtPlansUrl())) {
+            activity = new QuestionableActivityListingDTO();
+            activity.setBefore("Removed Plans URL: " + origListing.getRwtPlansUrl());
+        }
+        return activity;
+    }
+
+    public QuestionableActivityListingDTO checkRealWorldTestingResultsRemoved(
+            CertifiedProductSearchDetails origListing, CertifiedProductSearchDetails newListing) {
+
+        QuestionableActivityListingDTO activity = null;
+        if (!StringUtils.isEmpty(origListing.getRwtResultsUrl()) && StringUtils.isEmpty(newListing.getRwtResultsUrl())) {
+            activity = new QuestionableActivityListingDTO();
+            activity.setBefore("Removed Results URL: " + origListing.getRwtResultsUrl());
+        }
+        return activity;
+    }
+
     private Date getB3ChangeDate() throws ParseException {
         String dateAsString = env.getProperty(B3_CHANGE_DATE);
         SimpleDateFormat sdf = new SimpleDateFormat("MM/dd/yyyy");
@@ -476,7 +498,7 @@ public class ListingQuestionableActivityProvider {
         public int compare(CertificationStatusEvent a, CertificationStatusEvent b) {
             return a.getEventDate().longValue() < b.getEventDate().longValue()
                     ? -1
-                    : a.getEventDate().longValue() == b.getEventDate().longValue() ? 0 : 1;
+                            : a.getEventDate().longValue() == b.getEventDate().longValue() ? 0 : 1;
         }
     }
 }