+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/scheduler/job/DeveloperAccessReport.java
@@ -13,6 +13,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Locale;
 import java.util.Set;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 import javax.mail.MessagingException;
@@ -33,7 +34,7 @@ import org.springframework.web.context.support.SpringBeanAutowiringSupport;
 
 import gov.healthit.chpl.dao.CertificationBodyDAO;
 import gov.healthit.chpl.dao.impl.BaseDAOImpl;
-import gov.healthit.chpl.domain.Contact;
+import gov.healthit.chpl.domain.contact.PointOfContact;
 import gov.healthit.chpl.dto.CertificationBodyDTO;
 import gov.healthit.chpl.entity.CertificationStatusType;
 import gov.healthit.chpl.entity.auth.UserContactEntity;
@@ -182,7 +183,7 @@ public class DeveloperAccessReport extends QuartzJob {
         currRow.set(DEVELOPER_CONTACT_EMAIL, devAcbMap.getContactEmail());
         currRow.set(DEVELOPER_CONTACT_PHONE_NUMBER, devAcbMap.getContactPhoneNumber());
         currRow.set(DEVELOPER_USER_COUNT, developerAccessDao.getUserCountForDeveloper(devAcbMap.getDeveloperId()) + "");
-        List<Contact> userContactList = developerAccessDao.getContactForDeveloperUsers(devAcbMap.developerId);
+        List<PointOfContact> userContactList = developerAccessDao.getContactForDeveloperUsers(devAcbMap.developerId);
         currRow.set(DEVELOPER_USER_EMAILS, (userContactList == null) ? "" : formatContacts(userContactList));
         Date lastLoginDate = developerAccessDao.getLastLoginDateForDeveloper(devAcbMap.getDeveloperId());
         currRow.set(LAST_LOGIN_DATE, lastLoginDate == null ? "" : getTimestampFormatter().format(lastLoginDate));
@@ -207,7 +208,7 @@ public class DeveloperAccessReport extends QuartzJob {
         return row;
     }
 
-    private String formatContacts(List<Contact> userContactList) {
+    private String formatContacts(List<PointOfContact> userContactList) {
         List<String> contactStrings = new ArrayList<String>();
         userContactList.stream()
             .forEach(contact -> contactStrings.add(contact.getFullName() + " <" + contact.getEmail() + ">"));
@@ -224,14 +225,32 @@ public class DeveloperAccessReport extends QuartzJob {
     private void sendEmail(JobExecutionContext jobContext, List<List<String>> csvRows, List<CertificationBodyDTO> acbs)
             throws MessagingException {
         LOGGER.info("Sending email to {} with contents {} and a total of {} developer access rows",
-                getEmailRecipients(jobContext).get(0), getHtmlMessage(csvRows.size()), csvRows.size());
+                getEmailRecipients(jobContext).get(0), getHtmlMessage((csvRows.size()), getAcbNamesAsCommaSeparatedList(jobContext)));
 
         EmailBuilder emailBuilder = new EmailBuilder(env);
         emailBuilder.recipients(getEmailRecipients(jobContext))
                 .subject(getSubject(jobContext))
-                .htmlMessage(getHtmlMessage(csvRows.size()))
+                .htmlMessage(getHtmlMessage(csvRows.size(), getAcbNamesAsCommaSeparatedList(jobContext)))
                 .fileAttachments(getAttachments(csvRows, acbs))
                 .sendEmail();
+    }
+    
+    private String getAcbNamesAsCommaSeparatedList(JobExecutionContext jobContext) {
+        if (Objects.nonNull(jobContext.getMergedJobDataMap().getString("acb"))) {
+            return Arrays.asList(
+                    jobContext.getMergedJobDataMap().getString("acb").split(SchedulerManager.DATA_DELIMITER)).stream()
+                    .map(acbId -> {
+                        try {
+                            return certificationBodyDAO.getById(Long.parseLong(acbId)).getName();
+                        } catch (NumberFormatException | EntityRetrievalException e) {
+                            LOGGER.error("Could not retreive ACB name based on value: " + acbId, e);
+                            return "";
+                        }
+                    })
+                    .collect(Collectors.joining(", "));
+        } else {
+            return "";
+        }
     }
 
     private String getSubject(JobExecutionContext jobContext) {
@@ -258,11 +277,11 @@ public class DeveloperAccessReport extends QuartzJob {
         return csvFile;
     }
 
-    private String getHtmlMessage(Integer rowCount) {
+    private String getHtmlMessage(Integer rowCount, String acbList) {
         if (rowCount > 0) {
-            return String.format(env.getProperty("developerAccessHasDataEmailBody"), rowCount);
+            return String.format(env.getProperty("developerAccessHasDataEmailBody"), acbList, rowCount);
         } else {
-            return String.format(env.getProperty("developerAccessNoDataEmailBody"));
+            return String.format(env.getProperty("developerAccessNoDataEmailBody"), acbList);
         }
     }
 
@@ -325,8 +344,8 @@ public class DeveloperAccessReport extends QuartzJob {
         }
 
         @Transactional
-        public List<Contact> getContactForDeveloperUsers(Long developerId) {
-            List<Contact> contacts = new ArrayList<Contact>();
+        public List<PointOfContact> getContactForDeveloperUsers(Long developerId) {
+            List<PointOfContact> contacts = new ArrayList<PointOfContact>();
             Query query = entityManager.createQuery("SELECT contact "
                     + "FROM UserDeveloperMapEntity udm "
                     + "JOIN udm.developer developer "
@@ -345,7 +364,7 @@ public class DeveloperAccessReport extends QuartzJob {
                 return contacts;
             }
             for (UserContactEntity queryResult : queryResults) {
-                Contact contact = new Contact();
+                PointOfContact contact = new PointOfContact();
                 contact.setEmail(queryResult.getEmail());
                 contact.setFullName(queryResult.getFullName());
                 contacts.add(contact);